requirements.txt
beautifulsoup4==4.11.2
fastapi==0.112.2
google_search_results==2.4.2
langchain_core==0.2.37
langchain_openai==0.1.23
langgraph==0.2.16
pandas==2.2.2
pydantic==2.8.2
pymongo==4.3.3
python-decouple==3.8
python_dateutil==2.9.0.post0
Requests==2.32.3
serpapi==0.1.5
yfinance==0.2.41
